Rename the job to quicksuggest2bq
