+++ .circleci/config.yml
@@ -170,18 +170,18 @@ jobs:
           command: docker run app:build make lint
 
 
-  build-job-quicksuggest-2-bq:
+  build-job-quicksuggest2bq:
     docker:
       - image: docker:stable-git
     steps:
       - checkout
       - compare-branch:
-          pattern: ^jobs/quicksuggest-2-bq/
+          pattern: ^jobs/quicksuggest2bq/
       - setup_remote_docker:
           version: 19.03.13
       - run:
           name: Build Docker image
-          command: docker build -t app:build jobs/quicksuggest-2-bq/
+          command: docker build -t app:build jobs/quicksuggest2bq/
       - run:
           name: Test Code
           command: docker run app:build pytest --flake8 --black
@@ -296,16 +296,16 @@ workflows:
               only: main
 
 
-  job-quicksuggest-2-bq:
+  job-quicksuggest2bq:
     jobs:
-      - build-job-quicksuggest-2-bq
+      - build-job-quicksuggest2bq
       - gcp-gcr/build-and-push-image:
           context: data-eng-airflow-gcr
-          docker-context: jobs/quicksuggest-2-bq/
-          path: jobs/quicksuggest-2-bq/
-          image: quicksuggest-2-bq_docker_etl
+          docker-context: jobs/quicksuggest2bq/
+          path: jobs/quicksuggest2bq/
+          image: quicksuggest2bq_docker_etl
           requires:
-            - build-job-quicksuggest-2-bq
+            - build-job-quicksuggest2bq
           filters:
             branches:
               only: main
