--- jobs/quicksuggest-2-bq/ci_workflow.yaml
@@ -1,13 +1,0 @@
-job-quicksuggest-2-bq:
-  jobs:
-    - build-job-quicksuggest-2-bq
-    - gcp-gcr/build-and-push-image:
-        context: data-eng-airflow-gcr
-        docker-context: jobs/quicksuggest-2-bq/
-        path: jobs/quicksuggest-2-bq/
-        image: quicksuggest-2-bq_docker_etl
-        requires:
-          - build-job-quicksuggest-2-bq
-        filters:
-          branches:
-            only: main
