+++ jobs/quicksuggest2bq/ci_job.yaml
@@ -1,15 +1,15 @@
-build-job-quicksuggest-2-bq:
+build-job-quicksuggest2bq:
   docker:
     - image: docker:stable-git
   steps:
     - checkout
     - compare-branch:
-        pattern: ^jobs/quicksuggest-2-bq/
+        pattern: ^jobs/quicksuggest2bq/
     - setup_remote_docker:
         version: 19.03.13
     - run:
         name: Build Docker image
-        command: docker build -t app:build jobs/quicksuggest-2-bq/
+        command: docker build -t app:build jobs/quicksuggest2bq/
     - run:
         name: Test Code
         command: docker run app:build pytest --flake8 --black
